refactor(daemon): split memory/provider.ts into per-provider files
The 920-line provider.ts held five provider classes, the dispatcher and the
factories in one module. Move each into src/memory/providers/ — managed.ts,
none.ts, native.ts, external.ts, dispatching.ts, factory.ts — plus a shared
runtime-env.ts for the one helper managed/none/factory all use.

Pure moves: no behavior change. provider.ts keeps a re-export for every moved
symbol so its ~17 external importers stay untouched.
